[#noissue] Improved Reliability of PingSession lifecycle

+++ collector/src/main/java/com/navercorp/pinpoint/collector/service/async/AgentProperty.java
@@ -25,5 +25,5 @@ public interface AgentProperty {
     String getAgentId();
     long getStartTime();
     Object get(String key);
-    short getServiceType();
+    int getServiceType();
 }
